+++ src/component/signup/signupinputbox/SignupInputScreen.jsx
@@ -1,34 +1,99 @@
 import React from "react";
 import styled from "styled-components";
 import SocialLogin from "../sociallogin/SocialScreen";
-import IdInput from "./idinput/Screen";
-import PasswordInput from "./passwordinput/Screen";
-import NameInput from "./nameinput/Screen";
-import EmailInput from "./emailinput/Screen";
-import SignupBtn from "./signupbtn/SignupBtnScreen";
+import SignupBtn from "./SignupBtnScreen";
+import { useDispatch, useSelector } from "react-redux";
+import { changeField } from "../../../redux/modules/signUpSlice";
 
-const SignupInput = () => {
+const SignUpInput = () => {
+  const dispatch = useDispatch();
+  const { form } = useSelector(({ signUpPost }) => ({
+    form: signUpPost,
+  }));
+
+  const onChangeInputHandler = (event) => {
+    const { name, value } = event.target;
+    dispatch(
+      changeField({
+        key: name,
+        value,
+      })
+    );
+  };
+
   const img =
     "http://img.echosting.cafe24.com/skin/base/common/ico_required.gif";
   return (
     <SignupBox>
       <SocialLogin />
       <SignupDiv>
-        <IdInput starImg={img} />
+        <SignupDivSecond>
+          아이디
+          <img src={img}></img>
+        </SignupDivSecond>
+        <SignupIdInput
+          name="loginId"
+          onChange={onChangeInputHandler}
+        ></SignupIdInput>
+        {!/^(?=.*\d)(?=.*[a-z])[0-9a-z]{4,16}$/.test(form.loginId) &&
+          form.loginId && <SignupPtag></SignupPtag>}
+        <SignupPtag>(영문소문자/숫자,4~16자)</SignupPtag>
       </SignupDiv>
-      <PasswordInput starImg={img} />
       <SignupDiv>
-        <NameInput starImg={img} />
+        <SignupDivSecond>
+          비밀번호
+          <img src={img}></img>
+        </SignupDivSecond>
+
+        <SignupDivThird>
+          <SignupInput
+            name="password"
+            type="password"
+            onChange={onChangeInputHandler}
+          ></SignupInput>
+          <SignupPtagSecond>
+            (영문 대소문자/숫자/특수문자 중 3가지 이상 조합, 10자 ~ 16자)
+          </SignupPtagSecond>
+        </SignupDivThird>
       </SignupDiv>
       <SignupDiv>
-        <EmailInput starImg={img} />
+        <SignupDivSecond>
+          비밀번호 확인
+          <img src={img}></img>
+        </SignupDivSecond>
+        <SignupInput
+          name="passwordcheck"
+          type="password"
+          onChange={onChangeInputHandler}
+        ></SignupInput>
+        {form.password !== form.passwordcheck && form.passwordcheck !== "" && (
+          <SignupPtag>비밀번호가 일치하지 않습니다</SignupPtag>
+        )}
+      </SignupDiv>
+      <SignupDiv>
+        <SignupDivSecond>
+          이름
+          <img src={img}></img>
+        </SignupDivSecond>
+        <SignupInput name="name" onChange={onChangeInputHandler}></SignupInput>
+      </SignupDiv>
+      <SignupDiv>
+        <SignupDivSecond>
+          이메일
+          <img src={img}></img>
+        </SignupDivSecond>
+        <SignupInput
+          name="email"
+          type="email"
+          onChange={onChangeInputHandler}
+        ></SignupInput>
       </SignupDiv>
       <SignupBtn />
     </SignupBox>
   );
 };
 
-export default SignupInput;
+export default SignUpInput;
 
 const SignupBox = styled.div`
   display: flex;
@@ -47,4 +112,51 @@ const SignupDiv = styled.div`
   margin-top: 30px;
   padding-bottom: 30px;
   border-bottom: 1px solid var(--color-light-gray);
+`;
+
+const SignupDivThird = styled.div`
+  display: flex;
+  flex-direction: column;
+`;
+
+const SignupPtagSecond = styled.span`
+  font-size: 0.8rem;
+  margin-top: 10px;
+  padding-left: 70px;
+`;
+
+const SignupInput = styled.input`
+  float: left;
+  width: 250px;
+  height: 30px;
+  font-size: 1.2rem;
+  letter-spacing: 1px;
+  font-weight: bold;
+  padding: 2px 2px 2px 3px;
+  border: 1px solid var(--color-light-gray);
+  margin-left: 70px;
+  margin-right: 10px;
+  padding-right: 10px;
+`;
+
+const SignupDivSecond = styled.div`
+  width: 120px;
+`;
+
+const SignupPtag = styled.span`
+  font-size: 0.8rem;
+`;
+
+const SignupIdInput = styled.input`
+  float: left;
+  width: 250px;
+  height: 30px;
+  font-size: 1.2rem;
+  letter-spacing: 1px;
+  font-weight: bold;
+  padding: 2px 2px 2px 3px;
+  border: 1px solid var(--color-light-gray);
+  margin-left: 70px;
+  margin-right: 10px;
+  padding-right: 10px;
 `;
