+++ src/redux/modules/signUpSlice.jsx
@@ -1,24 +1,24 @@
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
-import { serverUrlApiV1 } from "../../core";
+import { instanceApiV1 } from "../../core/api";
 
 const initialState = {
-  signUp: {
-    loginId: "",
-    password: "",
-    passwordcheck: "",
-    name: "",
-    email: "",
-  },
-  isLoading: false,
-  result: null,
-  error: null,
+  loginId: "",
+  password: "",
+  passwordcheck: "",
+  name: "",
+  email: "",
 };
-
 export const __postSignUp = createAsyncThunk(
   "signUpPost/postSignUp",
   async (payload, thunkAPI) => {
+    const loginId = payload.loginId;
+    const password = payload.password;
+    const name = payload.name;
+    const email = payload.email;
+    const userInfo = { loginId, password, name, email };
     try {
-      const { data } = await serverUrlApiV1.post("/members/signup", payload);
+      const { data } = await instanceApiV1.post("/members/signup", userInfo);
+      console.log(data);
       return thunkAPI.fulfillWithValue(data);
     } catch (error) {
       alert("회원가입에 실패하였습니다");
@@ -31,26 +31,16 @@ const signUpSlice = createSlice({
   name: "signUpPost",
   initialState,
   reducers: {
-    changeField: (state, { payload: { form, key, value } }) => {
-      state[form][key] = value;
+    changeField: (state, { payload: { key, value } }) => {
+      state[key] = value;
     },
-    initializeForm: (state, { payload: { form } }) => ({
-      ...state,
-      [form]: initialState[form],
-    }),
   },
   extraReducers: (builder) => {
     builder
-      .addCase(__postSignUp.pending, (state) => {
-        state.isLoading = true;
-      })
       .addCase(__postSignUp.fulfilled, (state, action) => {
-        state.isLoading = false;
-        state.contents = action.payload;
         state.result = "success";
       })
       .addCase(__postSignUp.rejected, (state, action) => {
-        state.isLoading = false;
         state.error = action.payload;
       });
   },
